adding form buttons & handling submission
