+++ src/components/NewPost.jsx
@@ -1,16 +1,40 @@
 /* eslint-disable react/prop-types */
+import { useState } from 'react';
 import classes from './NewPost.module.css';
 
-function NewPost(props) {
+function NewPost({ onCancel }) {
+    const [enteredBody, setEnteredBody] = useState('');
+    const [enteredAuthor, setEnteredAuthor] = useState('');
+
+    function bodyChangeHandler(event) {
+        setEnteredBody(event.target.value)
+    }
+    function authorChangeHandler(event) {
+        setEnteredAuthor(event.target.value)
+    }
+    function handleSubmit(event) {
+        event.preventDefault();
+        const postData = {
+            author: enteredAuthor,
+            body: enteredBody
+        }
+        console.log(postData);
+        onCancel();
+    }
+
     return (
-        <form className={classes.form}>
+        <form className={classes.form} onSubmit={handleSubmit}>
             <p>
                 <label htmlFor="body">Text</label>
-                <textarea id="body" required rows={3} onChange={props.onBodyChange} />
+                <textarea id="body" required rows={3} onChange={bodyChangeHandler} />
             </p>
             <p>
                 <label htmlFor="name">Your name</label>
-                <input type="text" id="name" required onChange={props.onAuthorChange} />
+                <input type="text" id="name" required onChange={authorChangeHandler} />
+            </p>
+            <p className={classes.actions}>
+                <button type='button' onClick={onCancel}>Cancel</button>
+                <button>Save</button>
             </p>
         </form>
     );
